userbot/requirements.txt
pyrogram==2.0.106
tgcrypto
fastapi
uvicorn[standard]
python-dotenv
httpx
